Rename : Restructure folder and split main.py into routers, internal
